Refactor translations serializer into translations extractors
