+++ api/lib/infrastructure/translations-extractors/index.js
